Split model picker row chrome
