+++ docs/DECISIONS.md
@@ -146,7 +146,7 @@
 - Mixed command-palette search prioritizes concrete workspace actions over slash-command templates. Prefixing the query with `/` scopes the palette to slash templates, so users can discover command syntax without plain searches like "checks" or "approve pr" returning insert-only commands ahead of executable actions.
 - The sidebar's secondary tool surface is a single Tools disclosure plus a quiet Settings button. Terminal, Browser, Memories, Activity, and Command Palette are still one click away after opening Tools, but they no longer compete visually with the Codex-like primary navigation and thread list. Static HTML, SwiftUI, and Playwright share the same command IDs so this is a presentation simplification, not a second routing layer.
 - The native composer is a separate SwiftUI control instead of another private block inside `WorkspaceSwiftUIView`. Composer focus, send/stop actions, slash suggestions, and keyboard navigation now stay together, and the suggestion rows follow the shared polish contract: 40 pt minimum hit area, `0.96` press feedback, bounded command chips for long slash syntax, and quiet inline keyboard hints.
-- The native model picker is a separate SwiftUI control instead of a nested top-bar private block. The picker owns search focus, category sections, row expansion, favorite toggles, and model selection together, while approval mode stays in a neighboring composer control. Both controls keep 40 pt hit targets, shared `0.96` press feedback, and middle truncation for long provider/model metadata.
+- The native model picker is split by responsibility instead of living in the top bar or a single mixed picker file. `QuillCodeModelPickerView` owns the trigger, popover search text, search focus, keyboard highlight, empty/search states, and final model selection. `QuillCodeModelPickerRows.swift` owns category sections, option rows, favorite/detail actions, badges, expanded metadata, shared 40 pt hit targets, shared `0.96` press feedback, and middle truncation for long provider/model metadata. Approval mode stays in a neighboring composer control so model choice and autonomy stay separate.
 - The native top bar is a separate SwiftUI control. Workspace layout owns where the chrome sits, while `QuillCodeTopBarView` owns thread identity, runtime status, and the bounded utility overflow. Model/mode access lives with the composer so send-time controls do not crowd the title row. The ellipsis uses the shared 40 pt hit target and `0.96` press feedback so the quiet Codex-like chrome still feels tactile.
 - The native sidebar is a separate SwiftUI control. The workspace shell decides width and placement, while `QuillCodeSidebarView` owns primary navigation, thread sections, bulk selection, project rows, and the compact Tools/Settings footer. Sidebar commands still route through shared `WorkspaceCommandSurface` IDs, and high-frequency sidebar controls follow the same 40 pt hit target and `0.96` press feedback contract as other chrome.
 - The native review pane is a separate SwiftUI control. `QuillCodeReviewPaneView` owns review summary, file rows, hunk rows, inline comments, range notes, and review action buttons, while the workspace shell only decides when the review surface appears. Review controls use the shared 40 pt hit target and `0.96` press feedback contract so diff-review actions stay compact but tactile.
